feat: Github action for deploying workflows in examples (#855)
## Proposed changes

Ensure that we always deploy the workflows for examples when we merge.
Since some of these examples are going to be critical entrypoints for
new users, it is important that we ensure we don't accidentally deploy
breakages.

We will deploy the workflows to the `prod` environment, and ensure the
production versions of the examples on vercel use this environment.